Fixed warning messages in 64-bit

+++ qb.c
@@ -646,11 +646,11 @@ void qb_disable_zend_optimizer(TSRMLS_D) {
 		char *entry_name = "opcache.optimization_level";
 		char *entry_value = "0";
 #if !ZEND_ENGINE_2_2 && !ZEND_ENGINE_2_1
-		zend_alter_ini_entry_ex(entry_name, strlen(entry_name) + 1, entry_value, strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_RUNTIME, TRUE TSRMLS_CC);
+		zend_alter_ini_entry_ex(entry_name, strlen(entry_name) + 1, entry_value, (uint) strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_RUNTIME, TRUE TSRMLS_CC);
 #elif !ZEND_ENGINE_2_1
-		zend_alter_ini_entry_ex(entry_name, strlen(entry_name) + 1, entry_value, strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_RUNTIME, TRUE);
+		zend_alter_ini_entry_ex(entry_name, strlen(entry_name) + 1, entry_value, (uint) strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_RUNTIME, TRUE);
 #else
-		zend_alter_ini_entry(entry_name, strlen(entry_name) + 1, entry_value, strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_ACTIVATE);
+		zend_alter_ini_entry(entry_name, strlen(entry_name) + 1, entry_value, (uint) strlen(entry_value), PHP_INI_USER, PHP_INI_STAGE_ACTIVATE);
 #endif
 	}
 }
